+++ PotatoRaytracing/src/PotatoSphere.cs
@@ -1,19 +1,19 @@
 ﻿using System;
 using System.Drawing;
-using System.Numerics;
+using System.DoubleNumerics;
 
 namespace PotatoRaytracing
 {
     public class PotatoSphere : PotatoObject
     {
-        public readonly float Radius;
+        public readonly double Radius;
 
-        public PotatoSphere(Vector3 pos, float radius) : base(pos)
+        public PotatoSphere(Vector3 pos, double radius) : base(pos)
         {
             Radius = radius;
         }
 
-        public PotatoSphere(Vector3 pos, float radius, string texturePath = "Textures\\uvTexture.bmp") : base(pos, texturePath)
+        public PotatoSphere(Vector3 pos, double radius, string texturePath = "Textures\\uvTexture.bmp") : base(pos, texturePath)
         {
             Radius = radius;
         }
@@ -23,11 +23,12 @@ namespace PotatoRaytracing
             return GetUV(position.X, position.Y, position.Z, texture);
         }
 
-        public override Vector2 GetUV(float x, float y, float z, Bitmap texture)
+        public override Vector2 GetUV(double x, double y, double z, Bitmap texture)
         {
             double u = 0.5 - (Math.Atan2(z, x) * Constants.INV_DOUBLE_PI);
             double v = 0.5 - (Math.Asin(y) * Constants.INV_PI);
-            return new Vector2((float)u * texture.Width, (float)v * texture.Height);
+            
+            return new Vector2(u * texture.Width, v * texture.Height);
         }
     }
 }
